Users now don't need to show twice before returned data.tables print
Previously to view a returned data.table the user would have to type the name of the variable they've stored the data.table in and hit enter twice after running each function, and no output would be shown if the output of a function was not assigned to a variable. An empty show() call has been added on return so that now the output behaves as expected.